requirements.txt
kafka-python
requests
streamlit
numpy
pandas
tensorflow
scikit-learn
xlsxwriter